# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Get the current user
      responses:
        "200":
          description: The current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
    patch:
      operationId: updateMe
      tags: [me]
      summary: Update profile fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeBody"
      responses:
        "200":
          description: The updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
  /me/projects:
    get:
      operationId: listMyProjects
      tags: [me, projects]
      summary: List projects owned by the current user
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectListItem"
    post:
      operationId: createMyProject
      tags: [me, projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProjectResponse"
  /me/projects/{slug}:
    patch:
      operationId: renameMyProject
      tags: [me, projects]
      summary: Rename a project (and re-slug it)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameProjectBody"
      responses:
        "200":
          description: The updated project list row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListItem"
    delete:
      operationId: deleteMyProject
      tags: [me, projects]
      summary: Delete a project
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
...
  /projects/{username}/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by owner + slug
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{username}/{slug}/builds:
    get:
      operationId: listProjectBuilds
      tags: [projects, builds]
      summary: List builds for a project
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Builds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Build"
    post:
      operationId: createProjectBuild
      tags: [projects, builds]
      summary: Kick off a new build
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBuildBody"
      responses:
        "200":
          description: Created build
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Build"
...
  /projects/{username}/{slug}/domains/{host}:
    delete:
      operationId: removeProjectDomain
      tags: [projects, domains]
      summary: Remove a custom domain
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: host
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
...
  /explore:
    get:
      operationId: explore
      tags: [explore]
      summary: Explore public projects
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
        - name: framework
          in: query
          required: false
          schema: { type: string }
        - name: sort
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExploreItem"
...
components:
  schemas:
...
    Me:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        bio: { type: string }
        plan: { type: string }
        balance: { type: number }
        status: { type: string }
        signupDate:
          type: string
          description: YYYY-MM-DD signup date.
      required:
        - id
        - username
        - displayName
        - email
        - bio
        - plan
        - balance
        - status
        - signupDate
...
    UpdateMeBody:
      type: object
      properties:
        username: { type: string }
        displayName: { type: string }
        bio: { type: string }
...
    ProjectListItem:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        framework: { type: string }
        status: { type: string }
        isPublic: { type: boolean }
        clones: { type: integer }
        lastBuiltAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        buildsCount: { type: integer }
      required:
        - id
        - name
        - slug
        - description
        - framework
        - status
        - isPublic
        - clones
        - lastBuiltAt
        - buildsCount
    ProjectOwner:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        avatarUrl:
          type: string
          nullable: true
      required:
        - id
        - username
        - displayName
        - avatarUrl
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        framework: { type: string }
        status: { type: string }
        isPublic: { type: boolean }
        clones: { type: integer }
        createdAt:
          type: string
          format: date-time
        lastBuiltAt:
          type: string
          format: date-time
        owner:
          $ref: "#/components/schemas/ProjectOwner"
        buildsCount: { type: integer }
        lastBuildAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - name
        - slug
        - description
        - framework
        - status
        - isPublic
        - clones
        - createdAt
        - lastBuiltAt
        - owner
        - buildsCount
        - lastBuildAt
    Build:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        number: { type: integer }
        prompt: { type: string }
        aiMessage: { type: string }
        durationSec: { type: integer }
        cost: { type: number }
        filesChanged: { type: integer }
        tokensIn: { type: integer }
        tokensOut: { type: integer }
        model: { type: string }
        status: { type: string }
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - number
        - prompt
        - aiMessage
        - durationSec
        - cost
        - filesChanged
        - tokensIn
        - tokensOut
        - model
        - status
        - createdAt
    CreateBuildBody:
      type: object
      properties:
        prompt: { type: string }
        model: { type: string }
      required:
        - prompt
    CreateProjectBody:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        framework: { type: string }
      required:
        - name
    CreateProjectResponse:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        ownerUsername: { type: string }
      required:
        - id
        - slug
        - name
        - ownerUsername
    RenameProjectBody:
      type: object
      properties:
        name: { type: string }
      required:
        - name
    ExploreItem:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        framework: { type: string }
        clones: { type: integer }
        lastBuiltAt:
          type: string
          format: date-time
        author: { type: string }
        authorDisplayName: { type: string }
      required:
        - id
        - name
        - slug
        - description
        - framework
        - clones
        - lastBuiltAt
        - author
        - authorDisplayName
...