# lib/api-zod/src/generated/types/projectOwner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectOwner {
  id: string;
  username: string;
  displayName: string;
  avatarUrl: string | null;
}